Make TextPage an AnimatedSprite
Replace callback-driven letter-by-letter printing with
a method based on animation. Callbacks calls are synchronized
with the frame rate, meaning that callback-driven approach will
inevitably lead to the presence of time quants equal to 1 / fps.
Since delays between letters are intended to be very small, it
causes high inaccuracy in timing, which is not fatal, but still
harms UX. Animation-based approach is much more precise, since it
involves working with time intervals directly